Kairos.Types.TimeFormat

Description

Since: 0.1

Documentation

newtype TimeFormat

Time formatting string. The Monoid instance behaves like Text.

Since: 0.1

Constructors

MkTimeFormat

Fields

unTimeFormat :: Text

defaultTimeFormats :: NonEmpty TimeFormat

Parses 24 hour time w and wo colon (13:00, 1300) and 12 hour time w and wo passing (1:30 pm, 1 pm, 10:15 am).

Since: 0.1

hm :: TimeFormat

Format for 24-hour hours:minutes.

Since: 0.1

hm12h :: TimeFormat

Format for 12-hour hours:minutes am/pm.

Since: 0.1

hmTZ :: TimeFormat

Format for 24-hour hours:minutes TZ. As this contains a timezone flag, it should be used for formatting output only. In particular, it should not be used with format.

Since: 0.1

hmTZ12h :: TimeFormat

Format for 12-hour hours:minutes am/pm TZ. As this contains a timezone flag, it should be used for formatting output only. In particular, it should not be used with format.

Since: 0.1

rfc822 :: TimeFormat

Format for RFC822: %a, %_d %b %Y %H:%M:%S %Z. As this contains a timezone flag, it should be used for formatting output only. In particular, it should not be used with format.

Since: 0.1