Kairos.Types.TimeReader

Description

Since: 0.1

Documentation

data TimeReader

Determines how to read a time string.

Since: 0.1

Constructors

MkTimeReader

Fields

formats :: NonEmpty TimeFormat

Format(s) used when parsing the time string. This should not include timezone formatting e.g. %Z. Use srcTZ instead. It should also not include date information. Use date instead.

The formats are tried in order.

Since: 0.1

srcTZ :: Maybe TZInput

Timezone in which to read the string. Nothing corresponds to local timezone.

Since: 0.1

date :: Maybe Date

Date corresponding to the timeString. If Nothing, uses the current date, determined by the source.

Since: 0.1

timeString :: Text

The time string to parse. This should not include a timezone e.g. EST. Use srcTZ instead.

Since: 0.1

defaultTimeReader :: Text -> TimeReader

Given a time string, returns a default time reader.

Since: 0.1